Now checks for current version. Soon, used to skip
Almost doubled the size of the program, though most of that is in navigating Notepad to cull out the data I want. There's probably a faster, safer, and more invisible way to do that, but I don't know how yet.
Eventually I'll use this information to skip redundant installations, which will speed up the process in most cases and more importantly get rid of that dreadful Adobe Reader installation error it throws up when it's already installed.