Added basic flag support via files on desktop
I still haven't done anything with those 300 lines of version checking, but now I've added a kind of flash system. I couldn't find a way to directly read passed arguments or flags to a script, so I'm checking for files on the Desktop. For example, line 22 checks for a file called 'ahknorestart.txt' and if it exists, the script won't restart when finished. Now I can depreciate the spinoff script 'ua-sans-r', so that's cool.